move hovericon in asset to component
